refactor: hooksのディレクトリ修正 #184
